+++ packages/patrol_cli/lib/src/commands/develop_arg_parser.dart
@@ -25,7 +25,8 @@ void configureDevelopArgParser(PatrolCommand command) {
     ..usesAppNameOption()
     ..usesAndroidOptions()
     ..usesIOSOptions()
-    ..usesVideoRecordingOptions();
+    ..usesVideoRecordingOptions()
+    ..usesEmitTestManifestOption();
 
   command.argParser.addFlag(
     'open-devtools',